Bugfix: single argument topic not casting a list
Fixed a bug in PR #4 where if only a single topic is specified, the command does not send a tuple list in the payload, but only a string. Since the search_requester only handles topics and languages as lists, it loops through the single value printing each character instead of printing each topic item in the list. Now, the argument should always be cast as a list object